Create copy-text-into-clipboard.js
You can use it to copy a single word or a predefined text into the clipboard, using a simple JS function and an event call. 
Just add the function into your index.js, change the "texttobecopied" to your text and call it on the trigger. 
You can call it on buttons, links, images, texts or whatever you want to work as a trigger.